Move `GenreAmountProviderTests` to own file
